:recycle: [sendResettingEmailMessage] Move function into new service
