+++ src/ServiceV2/MailerService.php
@@ -10,7 +10,10 @@ use Psr\Log\LoggerInterface;
 use Symfony\Component\Mailer\Exception\TransportExceptionInterface;
 use Symfony\Component\Mailer\MailerInterface;
 use Symfony\Component\Mime\Email;
+use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
+use Symfony\Component\Routing\RouterInterface;
 use Symfony\Contracts\Translation\TranslatorInterface;
+use SymfonyCasts\Bundle\ResetPassword\Model\ResetPasswordToken;
 
 class MailerService
 {
@@ -21,6 +24,12 @@ class MailerService
         'en' => 3633387,
         'es' => 3633402,
     ];
+    private const RESET_PASSWORD_MAIL_TEMPLATE_ID = [
+        'fr' => 1553981,
+        'de' => 3616070,
+        'en' => 3633288,
+        'es' => 3633322,
+    ];
 
     /**
      * @param string[] $adminMails
@@ -29,6 +38,7 @@ class MailerService
         private readonly TranslatorInterface $translator,
         private readonly MailerInterface $mailer,
         private readonly LoggerInterface $logger,
+        private readonly RouterInterface $router,
         string $apikey,
         string $apisecret,
         private readonly string $noReplyMail,
@@ -96,5 +106,19 @@ class MailerService
                 $e->getMessage(),
             ));
         }
+    }
+
+    public function sendResettingEmailMessage(User $user, ResetPasswordToken $resetToken, string $locale = 'fr'): void
+    {
+        $templateId = self::RESET_PASSWORD_MAIL_TEMPLATE_ID[$locale] ?? self::RESET_PASSWORD_MAIL_TEMPLATE_ID['fr'];
+        $url = $this->router->generate('app_reset_password_email', [
+            'token' => $resetToken->getToken(),
+        ], UrlGeneratorInterface::ABSOLUTE_URL);
+        $variables = [
+            'SUBJECT_RESET_PASSWORD' => $this->translator->trans('mail_subject_reset_password'),
+            'MC_PREVIEW_TEXT' => '',
+            'RE_URL_RESET' => $url,
+        ];
+        $this->sendTemplate($templateId, $user->getEmail(), $variables);
     }
 }
